+++ lib/flutter_youtube.dart
@@ -5,52 +5,58 @@ import 'package:flutter/material.dart';
 
 class FlutterYoutube {
   static const MethodChannel _channel =
-      const MethodChannel('PonnamKarthik/flutter_youtube');
+      MethodChannel('PonnamKarthik/flutter_youtube');
 
   static const EventChannel _stream =
-      const EventChannel('PonnamKarthik/flutter_youtube_stream');
+      EventChannel('PonnamKarthik/flutter_youtube_stream');
 
   static String getIdFromUrl(String url, [bool trimWhitespaces = true]) {
-    if (url == null || url.length == 0) return null;
+    if (url == null || url.isEmpty) {
+      return null;
+    }
 
-    if (trimWhitespaces) url = url.trim();
+    if (trimWhitespaces) {
+      url = url.trim();
+    }
 
-    for (var exp in _regexps) {
-      Match match = exp.firstMatch(url);
-      if (match != null && match.groupCount >= 1) return match.group(1);
+    for (RegExp exp in _regexps) {
+      final Match match = exp.firstMatch(url);
+      if (match != null && match.groupCount >= 1) {
+        return match.group(1);
+      }
     }
 
     return null;
   }
 
-  static List<RegExp> _regexps = [
-    new RegExp(
-        r"^https:\/\/(?:www\.|m\.)?youtube\.com\/watch\?v=([_\-a-zA-Z0-9]{11}).*$"),
-    new RegExp(
-        r"^https:\/\/(?:www\.|m\.)?youtube(?:-nocookie)?\.com\/embed\/([_\-a-zA-Z0-9]{11}).*$"),
-    new RegExp(r"^https:\/\/youtu\.be\/([_\-a-zA-Z0-9]{11}).*$")
+  static final List<RegExp> _regexps = [
+    RegExp(
+        r'^https:\/\/(?:www\.|m\.)?youtube\.com\/watch\?v=([_\-a-zA-Z0-9]{11}).*$'),
+    RegExp(
+        r'^https:\/\/(?:www\.|m\.)?youtube(?:-nocookie)?\.com\/embed\/([_\-a-zA-Z0-9]{11}).*$'),
+    RegExp(r'^https:\/\/youtu\.be\/([_\-a-zA-Z0-9]{11}).*$')
   ];
 
-  static playYoutubeVideoByUrl(
-      {@required String apiKey,
-      @required String videoUrl,
-      bool autoPlay = false,
-      bool fullScreen = false,
-        Color appBarColor = Colors.black,
-        Color backgroundColor = Colors.black,
-      }) {
+  static playYoutubeVideoByUrl({
+    @required String apiKey,
+    @required String videoUrl,
+    bool autoPlay = false,
+    bool fullScreen = false,
+    Color appBarColor = Colors.black,
+    Color backgroundColor = Colors.black,
+  }) {
     if (apiKey.isEmpty || apiKey == null) {
-      throw "Invalid API Key";
+      throw 'Invalid API Key';
     }
 
     if (videoUrl == null || videoUrl.isEmpty) {
-      throw "Invalid Youtube URL";
+      throw 'Invalid Youtube URL';
     }
 
-    String id = getIdFromUrl(videoUrl);
+    final String id = getIdFromUrl(videoUrl);
 
     if (id == null) {
-      throw "Error extracting Youtube id from URL";
+      throw 'Error extracting Youtube id from URL';
     }
 
     final Map<String, dynamic> params = <String, dynamic>{
@@ -64,20 +70,20 @@ class FlutterYoutube {
     _channel.invokeMethod('playYoutubeVideo', params);
   }
 
-  static playYoutubeVideoById(
-      {@required String apiKey,
-      @required String videoId,
-      bool autoPlay = false,
-      bool fullScreen = false,
-      Color appBarColor = Colors.grey,
-      Color backgroundColor = Colors.black,
-      }) {
+  static playYoutubeVideoById({
+    @required String apiKey,
+    @required String videoId,
+    bool autoPlay = false,
+    bool fullScreen = false,
+    Color appBarColor = Colors.grey,
+    Color backgroundColor = Colors.black,
+  }) {
     if (apiKey.isEmpty || apiKey == null) {
-      throw "Invalid API Key";
+      throw 'Invalid API Key';
     }
 
     if (videoId == null || videoId.isEmpty) {
-      throw "Invalid Youtube URL";
+      throw 'Invalid Youtube URL';
     }
 
     final Map<String, dynamic> params = <String, dynamic>{
@@ -94,12 +100,12 @@ class FlutterYoutube {
   Stream<String> done;
 
   Stream<String> get onVideoEnded {
-    var d = _stream.receiveBroadcastStream().map<String>((element) => element);
+    final Stream<String> d =
+        _stream.receiveBroadcastStream().map<String>((element) => element);
     return d;
   }
 
-  int parseColorToHex(Color color){
-    return int.parse("0xFF"+color.value.toRadixString(16).padLeft(8, '0'));
+  static int parseColorToHex(Color color) {
+    return int.parse('0xFF' + color.value.toRadixString(16).padLeft(8, '0'));
   }
-
 }